+++ src/main/java/subway/ui/LineController.java
@@ -3,15 +3,11 @@ package subway.ui;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import subway.application.LineService;
-import subway.application.SectionService;
 import subway.dto.LineRequest;
 import subway.dto.LineResponse;
-import subway.dto.LineStationRequest;
-import subway.dto.LineStationResponse;
 
 import javax.validation.Valid;
 import java.net.URI;
-import java.sql.SQLException;
 import java.util.List;
 
 @RestController
@@ -19,11 +15,9 @@ import java.util.List;
 public class LineController {
 
     private final LineService lineService;
-    private final SectionService sectionService;
 
-    public LineController(LineService lineService, SectionService sectionService) {
+    public LineController(LineService lineService) {
         this.lineService = lineService;
-        this.sectionService = sectionService;
     }
 
     @PostMapping
@@ -51,29 +45,6 @@ public class LineController {
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteLine(@PathVariable Long id) {
         lineService.deleteLineById(id);
-        return ResponseEntity.noContent().build();
-    }
-
-    @ExceptionHandler(SQLException.class)
-    public ResponseEntity<Void> handleSQLException() {
-        return ResponseEntity.badRequest().build();
-    }
-
-    @PostMapping("/{lineId}/stations")
-    public ResponseEntity<Void> addStation(@PathVariable Long lineId, @RequestBody @Valid LineStationRequest lineStationRequest) {
-        Long id = sectionService.addStation(lineId, lineStationRequest);
-        return ResponseEntity.created(URI.create("/lines/" + lineId + "/" + id)).build();
-    }
-
-    @GetMapping("/{lineId}/stations")
-    public ResponseEntity<LineStationResponse> getLineStations(@PathVariable Long lineId) {
-        LineStationResponse lineStationResponse = sectionService.findByLineId(lineId);
-        return ResponseEntity.ok().body(lineStationResponse);
-    }
-
-    @DeleteMapping("/{lineId}/stations/{stationId}")
-    public ResponseEntity<Void> deleteStation(@PathVariable Long lineId, @PathVariable Long stationId) {
-        sectionService.removeStation(lineId, stationId);
         return ResponseEntity.noContent().build();
     }
 }
